Do not raise an exception on a faulty Auth Header
Currently if the token read from the header contains no spaces there will be an exception.
But this should only mean that you can't authorize with this header.
So I now check if the preg_split is successful and if no we just retun an empty
Permission Collection Transfer in the end.